requirements.txt
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
opencv-python==4.9.0.80
Pillow==10.3.0
streamlit==1.34.0
tqdm==4.66.4
mediapipe==0.10.20

